% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGarraDocker.R
\name{CreateSQLiteDataTable}
\alias{CreateSQLiteDataTable}
\title{Create DataBase SQLite Connection.}
\usage{
CreateSQLiteDataTable(dataTableName = "OGarraDockerDB", fieldsName)
}
\arguments{
\item{dataTableName}{Nome da tabela a fazer a conexão. Por padrão = "OGarraDockerDB"}

\item{fieldsName}{Nome das colunas a serem criadas para a tabela, pode ser no tipo de vector ou data frame}
}
\value{
Retorna verdadeiro caso foi possível a criação da tabela ou false, caso o contrário
}
\description{
Essa função cria uma tabela no tipo de sqlite

Essa função cria uma tabela no tipo de sqlite
}
\details{
* 1. Para obter informações a respeito do código R desta função, acessar ...
}
\examples{
CreateSQLiteDataTable(fieldsName = "id; nome; data"))

}
