% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Conexao com o BD SQL Server.}
\usage{
hello()
}
\arguments{
\item{driver}{driver SQL Server ODBC. Sys.getenv("CAEXDB_CONNECTION_DRIVER")="FreeTDS"}

\item{server}{server SQL Server. Default: Sys.getenv("CAEXDB_CONNECTION_SERVER")="caexsqldb.mppr"}

\item{port}{porta SQL Server. Default: Sys.getenv("CAEXDB_CONNECTION_PORT")="1433"}

\item{login}{login do usuário. Default: Sys.getenv("CAEXDB_CONNECTION_UID")}

\item{passwd}{senha do usuário. Default: Sys.getenv("CAEXDB_CONNECTION_PWD")}

\item{encod}{encoding. "UTF-8" para Linux e "latin1" para Windows}
}
\value{
Retorna uma lista de tabelas em formato dbplyr, que devem ser processadas pelas demais funções.
}
\description{
Esta função faz conexão com o BD SQL Server do NI a partir dos parâmetros informados. Como são setadas
variáveis de ambiente para os parâmetros, não é necessário fazer requisição desta função a partir da API OpenCPU. 
Esta função é utilizada de forma automática pelas demais funções. (Não utilizar!)
}
\details{
* 1. Para obter informações a respeito do código R desta função, acessar o link: http://opencpu.mppr/ocpu/library/BuscaContrato/R/SQLcon/print
}
\examples{
SQLcon()

}
