% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGarraDocker.R
\name{RemoveSQLiteDataTable}
\alias{RemoveSQLiteDataTable}
\title{Remove row from DataBase SQLite Connection.}
\usage{
RemoveSQLiteDataTable(dataTableName = "OGarraDockerDB", idToRemove)
}
\arguments{
\item{dataTableName}{Nome da tabela a fazer a conexão. Por padrão = "OGarraDockerDB"}

\item{idToRemove}{Id correspondente para remover o registro}
}
\value{
Retorna verdadeiro caso foi possível a remoção na tabela ou false, caso o contrário
}
\description{
Essa função remove registros de tabela no tipo de sqlite
}
\details{
* 1. Para obter informações a respeito do código R desta função, acessar ...
}
\examples{
RemoveSQLiteDataTable(idToRemove = 1)

}
