% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGarraDocker.R
\name{RetrieveRemoveSQLiteDataTable}
\alias{RetrieveRemoveSQLiteDataTable}
\title{Remove row from DataBase SQLite Connection and return the row that was removed.}
\usage{
RetrieveRemoveSQLiteDataTable(dataTableName = "OGarraDockerDB", idToRemove)
}
\arguments{
\item{dataTableName}{Nome da tabela a fazer a conexão. Por padrão = "OGarraDockerDB"}

\item{idToRemove}{Id correspondente para remover o registro}
}
\value{
Retorna o dado removido
}
\description{
Essa função remove registros de tabela no tipo de sqlite e a retorna para o chamamento
}
\details{
* 1. Para obter informações a respeito do código R desta função, acessar ...
}
\examples{
RemoveSQLiteDataTable(idToRemove = 1)

}
