% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGarraDocker.R
\name{RetrieveSQLiteData}
\alias{RetrieveSQLiteData}
\title{Retrieve data from DataBase SQLite Connection.}
\usage{
RetrieveSQLiteData(dataTableName = "OGarraDockerDB", idToRetrieve)
}
\arguments{
\item{dataTableName}{Nome da tabela a ser pego as informações. Por padrão = "OGarraDockerDB"}

\item{idToRetrieve}{Id correspondente para retirar do registro}
}
\value{
Retorna um json contendo as informações do banco pedidas
}
\description{
Essa função resgata os registros em que os ids forem iguais ao passado por parametro, deleta-os e retorna eles
}
\details{
* 1. Para obter informações a respeito do código R desta função, acessar ...
}
\examples{
GetAllSQLiteData(idToRetrieve = 1)

}
