% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGarraDocker.R
\name{util_remove_accents}
\alias{util_remove_accents}
\title{Remover Acentos}
\usage{
util_remove_accents(x = "")
}
\arguments{
\item{x}{string recebida normal}
}
\value{
Retorna a mesma string mas sem os acentos e tudo minusculo
}
\description{
Esta funcao remove acentos da string recebida e retorna a mesma string sem os mesmos e tudo minusculo.
}
\details{
* 1. Para obter informações a respeito do código R desta função, acessar o link: http://opencpu.mppr/ocpu/library/GeoEmpresas/R/util_remove_accents/print
}
\examples{
util_remove_accents("Coração")

}
