% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGarraDocker.R
\name{util_remove_special_characters}
\alias{util_remove_special_characters}
\title{Remover caracteres especiais}
\usage{
util_remove_special_characters(x = "")
}
\arguments{
\item{x}{string recebida com todos os caracteres}
}
\value{
Retorna a mesma string mas sem os caracteres especiais
}
\description{
Esta funcao remove caracteres especiais de string para conseguir processa-la de forma correta
}
\details{
* 1. Para obter informações a respeito do código R desta função, acessar o link: http://opencpu.mppr/ocpu/library/GeoEmpresas/R/util_remove_special_characters/print
}
\examples{
util_remove_special_characters("Avenida das Torres")
util_remove_special_characters("Avenida-das-Torres")

}
