% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OGarraDocker.R
\name{InsertSQLiteDataTable}
\alias{InsertSQLiteDataTable}
\title{Insert in DataBase SQLite Connection.}
\usage{
InsertSQLiteDataTable(dataTableName = "OGarraDockerDB", id, nome, data)
}
\arguments{
\item{dataTableName}{Nome da tabela a fazer a conexão. Por padrão = "OGarraDockerDB"}

\item{dataToInsert}{Data Frame contendo os dados a serem inseridos na tabela}
}
\value{
Retorna verdadeiro caso foi possível a insercao na tabela ou false, caso o contrário
}
\description{
Essa função insere registros em uma tabela no tipo de sqlite
}
\details{
* 1. Para obter informações a respeito do código R desta função, acessar ...
}
\examples{
InsertSQLiteDataTable(as.data.frame(list(id = c("1", "2", "3"), name = c("a", "b", "c"), data = c("12-10-2022", "10-4-1992", "25-4-2001"))))

}
